refactor redirects
