+++ frontend/auth/pages/index.tsx
@@ -1,16 +1,115 @@
+import { SelfServiceSettingsFlow } from '@ory/client'
 import type { NextPage } from 'next'
-import {
-  getInitialPropsRedirectIfLoggedOut,
-  NextPageContextWithSession,
-} from '../src/session'
+import { useEffect, useState } from 'react'
+import Flow from '../ui/Flow'
+import ory from '../src/ory'
+import { AxiosError } from 'axios'
+import { useSession } from '../src/session'
+import { useRouter } from 'next/router'
+import { handleFlowError } from '../src/errors'
+import MessagesList from '../ui/MessagesList'
 
-const Home: NextPage = () => {
-  return null
+interface Props {}
+
+const Settings: NextPage<Props> = () => {
+  const [flow, setFlow] = useState(
+    undefined as SelfServiceSettingsFlow | undefined,
+  )
+  const [session, setSession] = useSession()
+  const router = useRouter()
+  const { flow: flowId, return_to: returnTo } = router.query
+
+  useEffect(() => {
+    if (!session) {
+      router.replace('/login')
+      return
+    }
+
+    // Skip if we aren't ready
+    if (!router.isReady || flow) {
+      return
+    }
+
+    // If ?flow=.. was in the URL, we fetch it
+    if (flowId) {
+      ory
+        .getSelfServiceSettingsFlow(String(flowId))
+        .then(({ data }) => {
+          setFlow(data)
+        })
+        .catch(handleFlowError(router, 'settings', setFlow))
+      return
+    }
+
+    ory
+      .initializeSelfServiceSettingsFlowForBrowsers(
+        returnTo ? String(returnTo) : undefined,
+      )
+      .then(({ data }) => {
+        console.log(data)
+        setFlow(data)
+      })
+      .catch(handleFlowError(router, 'settings', setFlow))
+  }, [flowId, router, router.isReady, returnTo, flow])
+
+  if (!flow) {
+    return null
+  }
+
+  const onSubmit = async (data: any) => {
+    if (flow === undefined) {
+      console.error('no settings flow available to use')
+      return
+    }
+
+    await router.push(`/settings?flow=${flow?.id}`, undefined, {
+      shallow: true,
+    })
+
+    ory
+      .submitSelfServiceSettingsFlow(flow.id, data)
+      .then(async ({ data }) => {
+        console.log('Submitted settings flow', data)
+        setFlow(data)
+
+        // Update session with new data
+        const session = await ory.toSession()
+        setSession(session.data)
+      })
+      .catch(handleFlowError(router, 'settings', setFlow))
+      .catch(async (err: AxiosError) => {
+        // If the previous handler did not catch the error it's most likely a form validation error
+        if (err.response?.status === 400) {
+          setFlow(err.response.data as SelfServiceSettingsFlow)
+          return
+        }
+
+        debugger
+
+        return Promise.reject(err)
+      })
+  }
+
+  return (
+    <div>
+      <h1>Settings</h1>
+      <MessagesList messages={flow?.ui.messages} />
+      <h2>Profile</h2>
+      <Flow
+        flow={flow}
+        method="profile"
+        onSubmit={onSubmit}
+        hideGlobalMessages
+      />
+      <h2>Password</h2>
+      <Flow
+        flow={flow}
+        method="password"
+        onSubmit={onSubmit}
+        hideGlobalMessages
+      />
+    </div>
+  )
 }
 
-Home.getInitialProps = async (ctx: NextPageContextWithSession) => {
-  getInitialPropsRedirectIfLoggedOut(ctx)
-  return {}
-}
-
-export default Home
+export default Settings
